requirements.txt
asgiref==3.8.1
# backports.zoneinfo==0.2.1
chardet==5.2.0
Django==4.2.21
django-cors-headers==4.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
minimalmodbus==2.1.1
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pillow==10.4.0
PyMySQL==1.1.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
gunicorn==23.0.0
cryptography
'apache-airflow[celery,redis]'
python-dotenv==1.0.1
SQLAlchemy==1.4.52
